(Internal change: 1703730)

+++ pxr/base/lib/arch/hash.cpp
@@ -458,7 +458,7 @@ void SpookyHash::Hash128(
         while (u.p64 < end)
         { 
             Mix(u.p64, h0,h1,h2,h3,h4,h5,h6,h7,h8,h9,h10,h11);
-	    u.p64 += sc_numVars;
+            u.p64 += sc_numVars;
         }
     }
     else
@@ -467,7 +467,7 @@ void SpookyHash::Hash128(
         {
             memcpy(buf, u.p64, sc_blockSize);
             Mix(buf, h0,h1,h2,h3,h4,h5,h6,h7,h8,h9,h10,h11);
-	    u.p64 += sc_numVars;
+            u.p64 += sc_numVars;
         }
     }
 
@@ -475,7 +475,7 @@ void SpookyHash::Hash128(
     remainder = (length - ((const uint8 *)end-(const uint8 *)message));
     memcpy(buf, end, remainder);
     memset(((uint8 *)buf)+remainder, 0, sc_blockSize-remainder);
-    ((uint8 *)buf)[sc_blockSize-1] = remainder;
+    ((uint8 *)buf)[sc_blockSize-1] = static_cast<uint8>(remainder);
     
     // do some final mixing 
     End(buf, h0,h1,h2,h3,h4,h5,h6,h7,h8,h9,h10,h11);
@@ -563,7 +563,7 @@ void SpookyHash::Update(const void *message, size_t length)
         while (u.p64 < end)
         { 
             Mix(u.p64, h0,h1,h2,h3,h4,h5,h6,h7,h8,h9,h10,h11);
-	    u.p64 += sc_numVars;
+            u.p64 += sc_numVars;
         }
     }
     else
@@ -572,7 +572,7 @@ void SpookyHash::Update(const void *message, size_t length)
         { 
             memcpy(m_data, u.p8, sc_blockSize);
             Mix(m_data, h0,h1,h2,h3,h4,h5,h6,h7,h8,h9,h10,h11);
-	    u.p64 += sc_numVars;
+            u.p64 += sc_numVars;
         }
     }
